added utils and changed file name to monopoly
